PID Speed Compensation Initial Implementation

+++ Code/src/sysconfig.h
@@ -17,9 +17,10 @@ private:
 public:
     float duration  = 1;
     int   loopParam = 4;
-    int   timeStep  = 5e-3; 
+    float timeStep  = 5e-3; 
     int   resolution = 0; 
     float *calcedRPM;
+
     SysConfig(int IE_Ratio, int Resp_Rate, int Tidal_Volume);
     int get_IE_Ratio();
     int get_Resp_Rate();
